+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Big things planned! More unit tests! Less ORM quirks! More awesomeness!
 * **crud-controller:** add policies returned from getDestroyPolicies to the destroy route
 * **router:** allow policies to have their own policies; flatten duplicate policies
 * **crud-controller:** add policies returned from getQueryPolicies to the find and count routes
+* **server:** add startTime to the Server
 
 ### Bug Fixes